Move logging module
