+++ applications/lokqlDx-Av/Program.cs
@@ -4,6 +4,7 @@ using Avalonia.Platform.Storage;
 using HotAvalonia;
 using Intellisense.Configuration;
 using Jab;
+using LogSetup;
 using Lokql.Engine.Commands;
 using LokqlDx.Services;
 using LokqlDx.ViewModels;
