+++ libraries/LogSetup/LogSetup.csproj
@@ -7,6 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Jab" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" />
     <PackageReference Include="NLog"  />
   </ItemGroup>
