+++ test/IntellisenseTests/IntellisenseTests.csproj
@@ -30,6 +30,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\libraries\Intellisense\Intellisense.csproj" />
+    <ProjectReference Include="..\..\libraries\LogSetup\LogSetup.csproj" />
   </ItemGroup>
 
 </Project>
